move manual_control.py benchmark.py to examples
