+++ README.md
@@ -94,13 +94,13 @@ pip install -e .
 There is a UI application which allows you to manually control the agent with the arrow keys:
 
 ```
-./minigrid/manual_control.py
+./minigrid/examples/manual_control.py
 ```
 
 The environment being run can be selected with the `--env` option, eg:
 
 ```
-./minigrid/manual_control.py --env MiniGrid-Empty-8x8-v0
+./minigrid/examples/manual_control.py --env MiniGrid-Empty-8x8-v0
 ```
 
 ## Reinforcement Learning
