+++ minigrid/examples/manual_control.py
