+++ tests/test_scripts.py
@@ -1,8 +1,8 @@
 import gym
 import numpy as np
 
-from minigrid.benchmark import benchmark
-from minigrid.manual_control import key_handler, reset
+from minigrid.examples.benchmark import benchmark
+from minigrid.examples.manual_control import key_handler, reset
 from minigrid.window import Window
